Refactor calls interfaces and add https
